+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/MatrixClientConfiguration.kt
@@ -179,6 +179,7 @@ data class MatrixClientConfiguration(
         val roomOutboxMessage: Duration,
         val room: Duration,
         val presence: Duration,
+        val media: Duration,
     ) {
 
 
@@ -205,6 +206,7 @@ data class MatrixClientConfiguration(
                     roomOutboxMessage = duration / 2,
                     room = duration * 10,
                     presence = duration,
+                    media = duration / 6
                 )
         }
     }