+++ trixnity-utils/src/commonMain/kotlin/net/folivo/trixnity/utils/ByteArrayFlow.kt
@@ -10,6 +10,7 @@ fun ByteArray.toByteArrayFlow(): ByteArrayFlow = flowOf(this)
 
 suspend fun ByteArrayFlow.toByteArray(): ByteArray {
     val allByteArrays = toList()
+    if (allByteArrays.size == 1) return allByteArrays.first()
     val concatByteArray = ByteArray(allByteArrays.sumOf { it.size })
     var byteArrayPosition = 0
     allByteArrays.forEach { byteArray ->
@@ -21,6 +22,29 @@ suspend fun ByteArrayFlow.toByteArray(): ByteArray {
     return concatByteArray
 }
 
+/**
+ * Returns null, when [maxSize] exceeded.
+ */
+suspend fun ByteArrayFlow.toByteArray(maxSize: Long): ByteArray? =
+    try {
+        limitedSize(maxSize).toByteArray()
+    } catch (e: Exception) {
+        if (e is MaxByteArrayFlowSizeException || e.cause is MaxByteArrayFlowSizeException) null
+        else throw e
+    }
+
+private fun ByteArrayFlow.limitedSize(maxSize: Long): ByteArrayFlow = flow {
+    var size = 0
+    collect { nextBytes ->
+        size += nextBytes.size
+        if (size > maxSize) throw MaxByteArrayFlowSizeException
+        else emit(nextBytes)
+    }
+}
+
+private object MaxByteArrayFlowSizeException : RuntimeException()
+
+@Deprecated("use case missing, use toByteArray(maxSize: Long) instead")
 fun ByteArrayFlow.takeBytes(size: Int): ByteArrayFlow = flow {
     var currentSize = 0
     takeWhile { next ->